### 2019-08-16 #1653 @mlindsk
I have looked for information on the use of  external latex packages in paper.md but failed. Is it not possible to include such packages as the latex algorithm package?
