### 2019-08-16 #1653 @arfon
> @danielskatz I have looked for information on the use of external latex packages in paper.md but failed. Is it not possible to include such packages as the latex algorithm package?

We generally don't support custom packages with the paper build. Can you tell me what you're trying to achieve here, and could it be accomplished with regular (LaTeX) maths syntax?
