### 2019-08-16 #1653 @mlindsk
I am trying to display a pseudo algorithm in the algorithmics environment which cannot be done in regular latex. If it's a hassle I can certainly find other ways to achieve something similar. I just like the format.
