### 2019-08-17 #1653 @mlindsk
Yes sure. I will try to make a standalone tex file for the algorithm and convert to png.
