+++ README.Rmd
@@ -25,38 +25,38 @@ An **R** package to perform outlier detection in contingency tables using decomp
 ## Getting Started
 If you want to learn the "behind the scenes" of the model it is recommended to go through [The Outlier Model](https://mlindsk.github.io/molic/articles/) and look at the [documentation](https://mlindsk.github.io/molic/reference/index.html) as you read along. See also the examples below and the paper [TBA].
 
-## Installation
 You can install the development version of the package by using the `devtools` package:
 
 ```{r, eval = FALSE}
-devtools::install_github("mlindsk/molic", build_vignettes = TRUE)
+devtools::install_github("mlindsk/molic", build_vignettes = FALSE)
 ```
 
-## Main Functions
-A list of some core functions in the **molic** package is listed below
+## How To Cite
 
-```{r molic_funcs, echo = FALSE}
-func_names <- c(
-  "`outlier_model`",
-  "`p_val`",
-  "`efs`",
-  "`cl_tree`")
-func_descr <- c(
-  "Fits an outlier detection model",
-  "Calculates the p-value of an observation being an outlier",
-  "Fits a decomposable graph using forward selection.",
-  "Fits a tree (decomposable graph) using the Chow-Liu method."
-)
-func_tab <- data.frame(Function = func_names, Description = func_descr)
-pander::panderOptions('table.alignment.default', 'left')
-pander::pander(func_tab)
+ - If you want to cite the **outlier method** please use
+
+```latex
+@article{key,
+  title={Outlier Detection in Contingency Tables using Decomposable Graphical Models},
+  author={Lindskou, Mads and Eriksen, Poul Svante and Tvedebrink, Torben},
+  journal={Scandinavian Journal of Statistics},
+  doi={10.1111/sjos.12407},
+  year={2019}
+}
 ```
+
+ - If you want to cite the **molic** package please use
+
+```latex
+TBA
+```
+
 
 ## Example - Outlier Detection
 
 To demonstrate the outlier method we use the `car` data set from the [UCI Machine Learning Repository](https://archive.ics.uci.edu/ml/index.php). The data have $4$ classes that labels the evaluation of a car; `unacceptable, acc, vgood` and `good`. These classes are determined by the other variables in the data - and theses are _not_ necessarily independent of each other and we must therefore "fit their association".
 
-### Reading data
+### Reading Data
 
 ```{r reading_car, message = FALSE, warning = FALSE}
 library(dplyr)
@@ -68,7 +68,7 @@ car <- read.table("https://archive.ics.uci.edu/ml/machine-learning-databases/car
 colnames(car) <- c("buying", "maint", "doors", "persons", "lug", "safety", "class")
 ```
 
-### Defining subclasses
+### Defining Sub-Classes
 
 ```{r vgood_uacc}
 vgood_cars <- car %>%
@@ -82,7 +82,7 @@ unacc_cars <- car %>%
   mutate_all(.funs = function(x) substr(x, 1, 1))
 ```
 
-### Fitting an interaction graph
+### Fitting an Interaction Graph
 
 ```{r echo = FALSE, message = FALSE, warning = FALSE}
 library(molic)
@@ -104,7 +104,7 @@ plot(G_unacc)
 
 It is apparent that very good cars and unacceptable cars are determined by two different mechanisms.
 
-### Outlier test
+### Outlier Test
 
 We randomly select five cars from the `unacc_cars` data and test if they are outliers in `vgood_cars`.
 
@@ -142,6 +142,3 @@ So the class of a car is actually determined by all variables except for `doors`
 adj_list(G_car)$class
 ```
 We can also state e.g. that the `safety` of a car is independent of the price (the `buying` varible) when the class of the car is known; this phenomena is also known as _conditional independence_. 
-
-## How To Cite
-TBA
