### 2019-08-17 ddcc546
removed algorithm package
