### 2019-08-18 #1653 @csoneson
👋mlindsk - I will handle your submission. Do you have suggestions for suitable reviewers (e.g. from the list linked in the first comment above)?

A couple of initial comments, in addition to @danielskatz's above:
- [ ] Please only include the references that you are citing in the .bib file
- [ ] It looks like there are no actual tests in the `testthat/` folder - could you add that?
- [ ] Check the capitalization of "Proceedings of the seventeenth conference..."
