### 2019-08-19 45172bd
updated paper and more tests
