+++ paper/paper.bib
@@ -9,25 +9,10 @@
 @inproceedings{deshpande01_efficient,
   title={Efficient stepwise selection in decomposable models},
   author={Deshpande, Amol and Garofalakis, Minos and Jordan, Michael I},
-  booktitle={Proceedings of the Seventeenth conference on Uncertainty in artificial intelligence},
+  booktitle={Proceedings of the Seventeenth Conference on Uncertainty in Artificial Intelligence},
   pages={128--135},
   year={2001},
   organization={Morgan Kaufmann Publishers Inc.}
-}
-
-@book{lauritzen1996graphical,
-  title={Graphical models},
-  author={Lauritzen, Steffen L},
-  volume={17},
-  year={1996},
-  publisher={Clarendon Press}
-}
-
-@book{whittaker2009graphical,
-  title={Graphical models in applied multivariate statistics},
-  author={Whittaker, Joe},
-  year={2009},
-  publisher={Wiley Publishing}
 }
 
 @article{10002015global,
@@ -39,4 +24,24 @@
   pages={68},
   year={2015},
   publisher={Nature Publishing Group}
+}
+
+@Article{tvedebrink2018aims,
+  author = 	 {T Tvedebrink and P S Eriksen and H S Mogensen and
+                  N Morling},
+  title = 	 {{Weight of the Evidence of Genetic Investigations
+    of Ancestry Informative Markers}},
+  journal = 	 {Theoretical Population Biology},
+  year = 	 2018,
+  volume = 	 120,
+  pages = 	 {1-10}}
+
+
+@article{kidd2014microhaplo,
+title = {Current sequencing technology makes microhaplotypes a powerful new type of genetic marker for forensics},
+journal = {Forensic Science International: Genetics},
+volume = 12,
+pages = {215 - 224},
+year = 2014,
+author = {Kenneth K. Kidd and Andrew J. Pakstis and William C. Speed and Robert Lagacé and Joseph Chang and Sharon Wootton and Eva Haigh and Judith R. Kidd}
 }
