### 2019-08-19 #1653 @mlindsk
> wavemlindsk - I will handle your submission. Do you have suggestions for suitable reviewers (e.g. from the list linked in the first comment above)?
>
> A couple of initial comments, in addition to @danielskatz's above:
>
>     * [ x ]  Please only include the references that you are citing in the .bib file
>
>     * [ x ]  It looks like there are no actual tests in the `testthat/` folder - could you add that?
>
>     * [ x ]  Check the capitalization of "Proceedings of the seventeenth conference..."

:wave: @csoneson Updates according to your comments.
