### 2019-08-19 #1653 @csoneson
@mlindsk - great, thanks. While we are waiting to see whether @VivekTodur, @jdeligt or @jkanche would be available and interested in reviewing the submission, a couple of more things for the paper that can just as well be fixed now:
- please add DOIs to all your references wherever possible
- could you fix the "Consortium, 1. G. P. & others" in the reference list?
- when a citation is part of a sentence, please give it without the enclosing square brackets in the markdown file (e.g. in the first sentence of the second paragraph of the use case, but there are also other instances).
- I still think that "Proceedings of the seventeenth..." should have some words capitalized (e.g. according to https://dl.acm.org/citation.cfm?id=647235&picked=prox)
