### 2019-08-20 #1653 @mlindsk
@csoneson Goodmorning.

>     * please add DOIs to all your references wherever possible
 - check
>     * could you fix the "Consortium, 1. G. P. & others" in the reference list?
 - It renders fine when i'm running pandoc locally - i dont have control over the JOSS config. At least I wouldn't know how to change this. Any advice would be helpful.
>     * when a citation is part of a sentence, please give it without the enclosing square brackets in the markdown file (e.g. in the first sentence of the second paragraph of the use case, but there are also other instances).
 - Thanks - I was not aware that I could simply remove the brackets.
>     * I still think that "Proceedings of the seventeenth..." should have some words capitalized (e.g. according to https://dl.acm.org/citation.cfm?id=647235&picked=prox)
 - I have capitalized exactly the same words as in the reference. Again, it is something the JOSS pandoc config does.
