### 2019-08-20 #1653 @csoneson
I think both of these issues can be solved by an extra pair of curly brackets. I just sent a PR.
