### 2019-08-20 #1653 @mlindsk
@csoneson
> I think both of these issues can be solved by an extra pair of curly brackets. I just sent a PR.

Fantastic to learn new things in bibfiles. Seems corrected now.

@jdeligt
> Count me in

Fantastic!
