### 2019-08-20 #1665 @whedon
[REVIEW]: molic: An R package for multivariate outlier detection in contingency tables
**Submitting author:** @mlindsk (<a href="http://orcid.org/0000-0002-1033-697X">Mads Lindskou</a>)
**Repository:** <a href="https://github.com/mlindsk/molic" target ="_blank">https://github.com/mlindsk/molic</a>
**Version:** v1.0.0
