### 2019-08-21 mlindsk/molic#11 @jdeligt
Please add version and release
Related to https://github.com/openjournals/joss-reviews/issues/1665 checklist.
I can't find any releases in the repo nor a version in the main .R source file.
Please add these.
