### 2019-08-21 #1665 @jdeligt
@mlindsk could you please confirm when you have gone through the checklist. I can't find a version/release; https://github.com/mlindsk/molic/issues/11.
