+++ DESCRIPTION
@@ -1,6 +1,6 @@
 Package: molic
 Title: Multivariate Oulier Detection In Contingency Tables
-Version: 0.5.0
+Version: 1.0.0
 Authors@R: 
     person(given = "Mads",
            family = "Lindskou",
