### 2019-08-21 #1665 @mlindsk
@jdeligt I have now gone through the list and I have made a release v1.0.0.
