### 2019-08-27 #1665 @jdeligt
@mlindsk I'll progress with the testing of the software but please update the version of the publication (v0.5.0) to match the github release
