### 2019-08-27 #1665 @mlindsk
@csoneson Could you change the version as requested to v1.0.0. Thank you.

> @mlindsk I'll progress with the testing of the software but please update the version of the publication (v0.5.0) to match the github release
