### 2019-08-27 #1665 @jdeligt
Hi @mlindsk, some small comments first; while you're on it please also update the version in your documentation to match. I also can't find a "Contributing" doc or text about community guidelines in the README. Could the proceeding reference point to this: https://arxiv.org/abs/1301.2267 to have a stable identifier?

Then a slightly bigger comment, I'm making some assumptions about what you mean please correct me if they are incorrect. To me both the example data and the suggested Genetics seem 'perfect world' examples. i.e. the 2 datasets are very clearly distinguishable (i.e. the minigaplotypes have been selected to distinguish these groups) and as such they are confidently identified as outliers. How does the algorithm perform on an individual that has mixed haplotypes? And something that I'm also wondering about (but this might be my unfamiliarity with the specific forensic implications), why are traditional approaches like phylogenetics / Hemming distance / PCA / kmers / ... not useful to distinguish the groups? I can see how the algorithm could offer a generalized approach to outlier detection for non-categorical data but I struggle to see the added value for the genetics case.
