### 2019-08-27 #1665 @mlindsk
Hi @jdeligt

Version should now be updated (hopefully all places now). I am not sure, that adding a contributing doc would add much. Would it suffice if (in README) I include a section "Contribute, issues, and support" as done at https://github.com/mikldk/malan ?

According to the reference. Are you asking if I could skip the reference to [https://doi.org/10.1109/ictai.2004.100](url) ? If so; it adds a crucial mistake from [https://arxiv.org/abs/1301.2267](url) and cannot be left out. They complement each other.

It is true, that Europeans and Africans are easy to distinguish based on these haplotypes. However, it is not that easy for populations that are more "similar" as seen in Table 1 in the paper.md. Even though the haplotypes are selected for discriminatory purposes, we still need a (probabilistic) model to decide whether or not a profile is an outlier. The outlier model is only "as good as data" allow. What the model adds, is the possibility to take advantages of the mutual relationships among all variables. If one falsely assume independence between all variables the results are much worse (see Table 1).

Our method is an outlier test, which the models you refer to are not. Moreover, our model has the ability to investigate why a profile was declared an outlier since we have this nice graphical representation.

The method is really not for distinguishing groups but for outlier detection - a profile might not be declared as an outlier in several populations.

Within the genetic world our method is a direct generalization of the community accepted method [https://www.fsigeneticssup.com/article/S1875-1768(17)30216-0/fulltext](url) which only handles independent SNPs (a graph with no edges)

Also the software is intended for completely general purposes. The only requirement is data is categorical.

I hope this clarifies things - just ping me again if something is unclear.
