### 2019-08-27 #1665 @jdeligt
Hi @mlindsk, I'll try to clarify and answer the questions in order.

1. (Contributing) That would be the minimal thing to do, it helps people in understanding what model you prefer and what to expect. Some guidelines/examples are here: https://help.github.com/en/articles/setting-guidelines-for-repository-contributors

2. (Reference) I was refeering to "deshpande01_efficient" which at the moment does not have a DOI in paper.bib

3. (Genomics) I'm not debating wether the algorithm is a usefull outlier detection tool. I'm debating 2 things, a. Do the provided examples sufficiently show that it is a sensitive / usefull outlier detection tool for complex data. b. Is the biological use case you describe apporiate to substantiate your claim of it being usefull (I'll ellaborate a bit more below).

To prefent the review from becoming a clah of fields I would suggest adding some text to clarify why this method is relevant for this problem within the particular setting and to put it in context of the wider population genetics field. And please review wether the comparison you make is fair/valid when taking into acocunt the biological constraints that informed the data selection. You could consider to just hava a statement about how this approach properly accounts for the underlying biology,

(Genomics extened) The loci you study have been selected to be in LD and therfore linked, this means that biology dictates that performance will be worse if not considered as such. Therefor an algorithm that takes this into account will always do better than one that is not. Based on the sole fact that it approriatly accounts for the relationships. This means that your algorithm is the apporiate choice for this problem, not that the performace is better. The other approach is simply not fit to do this. Therefore the comparison (to mee) is mute.

(Ancestry extended) As indicated I cannot speak from a forencics background but this method is not the accepted standerd in human genomics and population genetics (a few examples; https://www.nature.com/articles/nature18964, https://www.ncbi.nlm.nih.gov/pmc/articles/PMC4836868/, https://www.nature.com/articles/s41586-018-0579-z). This makes it difficult for me to see why an outlier detection would be the approriate choice for forencics. I.e. I would expect that forencics would need an assignment of each haplotype to its respective ancestry to support identification. I.e. a mixed haplotype being an outlier for all populations does not seem the most usefull (again I can be horribly wrong but from a biology perspective this would be what I would expect).

Hope this helps to clarify where I'm coming from and provide some more context to the suggested changes to the 'Statement of need'.

All the best, Joep
