### 2019-08-28 #1665 @mlindsk
> Hi @mlindsk, I'll try to clarify and answer the questions in order.
>
>     1. (Contributing) That would be the minimal thing to do, it helps people in understanding what model you prefer and what to expect. Some guidelines/examples are here: https://help.github.com/en/articles/setting-guidelines-for-repository-contributors

Ok. Agreed.

>
>     2. (Reference) I was refeering to "deshpande01_efficient" which at the moment does not have a DOI in paper.bib

Ok.

>
>     3. (Genomics) I'm not debating wether the algorithm is a usefull outlier detection tool. I'm debating 2 things, a. Do the provided examples sufficiently show that it is a sensitive / usefull outlier detection tool for complex data. b. Is the biological use case you describe apporiate to substantiate your claim of it being usefull (I'll ellaborate a bit more below).
>

We agree that data could be more complex. However, increasing the complexity in the examples does not necessarily help the reader in understanding how the software works.

>
> To prefent the review from becoming a clah of fields I would suggest adding some text to clarify why this method is relevant for this problem within the particular setting and to put it in context of the wider population genetics field. And please review wether the comparison you make is fair/valid when taking into acocunt the biological constraints that informed the data selection. You could consider to just hava a statement about how this approach properly accounts for the underlying biology,
>

The DGMs are fitted separately within in each population and haplotype. For a specific population, the fitted DGMs are merged to a union graph. Thus, we take into account the biological nature. I.e. we exploit that haplotypes are independent between chromosomes.

> (Genomics extened) The loci you study have been selected to be in LD and therfore linked, this means that biology dictates that performance will be worse if not considered as such. Therefor an algorithm that takes this into account will always do better than one that is not. Based on the sole fact that it approriatly accounts for the relationships. This means that your algorithm is the apporiate choice for this problem, not that the performace is better. The other approach is simply not fit to do this. Therefore the comparison (to mee) is mute.
>

**(1)**

Yes, our expectation was indeed that the model would outperform the one neglecting dependencies. But the extend to which it outperforms the model with independencies is not obvious before doing the tests. Also, currently there are no other outlier detection tool using decomposable graphical models (DGMs). And the use of DGMs allow for a solid and well understood statistical way for reading off conditional independencies among the variables. One model that we can compare to though is the saturated model (a complete graph) corresponding to using frequency estimates. However, for high-dimensional data the saturated model breaks down since you would need an impossible amount of observations to catch rare events. Our approach remedies this, by focusing on local information (cliques in the fitted graph) from which it is way easier to investigate the nature of an outlier.

> (Ancestry extended) As indicated I cannot speak from a forencics background but this method is not the accepted standerd in human genomics and population genetics (a few examples; https://www.nature.com/articles/nature18964, https://www.ncbi.nlm.nih.gov/pmc/articles/PMC4836868/, https://www.nature.com/articles/s41586-018-0579-z). This makes it difficult for me to see why an outlier detection would be the approriate choice for forencics. I.e. I would expect that forencics would need an assignment of each haplotype to its respective ancestry to support identification. I.e. a mixed haplotype being an outlier for all populations does not seem the most usefull (again I can be horribly wrong but from a biology perspective this would be what I would expect).
>

**(2)**

What you refer to here is _population structure_ which is _not_ the same thing as an outlier test. In the forensic world, one could reach out for the STRUCTURE algorithm (https://www.ncbi.nlm.nih.gov/pubmed/10835412) for markers not in LD e.g. The problem here is, that if the (unknown) profile does _not_ originate from any of the populations/components in the database, it will be assigned to the most probable one (under the model) which is simply wrong. In a court (forensic science), it is then better to state that the profile can not be assumed to originate from any of the populations in the reference database. To put it another way; we may have _exclusive_ sample populations, but we do not have an _exhaustive_ database of reference populations. This statement is crucial.

I have added the last part of **(1)** and a "Expert Knowledge" section paper.md to clarify things. Let me know if I have completely misunderstood you.

> Hope this helps to clarify where I'm coming from and provide some more context to the suggested changes to the 'Statement of need'.
>

I would like to thank you for the time spend to reflect on the genetic applications. I hope, that we converge towards a common understanding all though it is a difficult task for two different backgrounds.

Regards,
Mads
