### 2019-09-01 #1665 @jkanche
@mlindsk devtools::check fails for the package.

Here's the error -

```
E  creating vignettes (33s)
   --- re-building ‘outlier_model.Rmd’ using rmarkdown

   Attaching package: 'dplyr'

   The following objects are masked from 'package:stats':

       filter, lag

   The following objects are masked from 'package:base':

       intersect, setdiff, setequal, union

   Quitting from lines 48-62 (outlier_model.Rmd)
   Error: processing vignette 'outlier_model.Rmd' failed with diagnostics:
   there is no package called 'pander'
   --- failed re-building ‘outlier_model.Rmd’

   SUMMARY: processing the following file failed:
     ‘outlier_model.Rmd’

   Error: Vignette re-building failed.
   Execution halted
Error in (function (command = NULL, args = character(), error_on_status = TRUE,  :
  System command error
Calls: <Anonymous> ... with_envvar -> force -> do.call -> <Anonymous> -> throw
Execution halted

Exited with status 1.
```

here's my sessionInfo

```
> sessionInfo()
R version 3.6.1 (2019-07-05)
Platform: x86_64-pc-linux-gnu (64-bit)
Running under: Manjaro Linux

Matrix products: default
BLAS:   /usr/lib/libopenblasp-r0.3.6.so
LAPACK: /usr/lib/liblapack.so.3.8.0

locale:
 [1] LC_CTYPE=en_US.UTF-8       LC_NUMERIC=C
 [3] LC_TIME=en_US.UTF-8        LC_COLLATE=en_US.UTF-8
 [5] LC_MONETARY=en_US.UTF-8    LC_MESSAGES=en_US.UTF-8
 [7] LC_PAPER=en_US.UTF-8       LC_NAME=C
 [9] LC_ADDRESS=C               LC_TELEPHONE=C
[11] LC_MEASUREMENT=en_US.UTF-8 LC_IDENTIFICATION=C

attached base packages:

other attached packages:
[1] dplyr_0.8.3 molic_1.0.0

 [1] Rcpp_1.0.1       lattice_0.20-38  crayon_1.3.4     assertthat_0.2.1
 [5] grid_3.6.1       R6_2.4.0         magrittr_1.5     pillar_1.4.2
 [9] rlang_0.4.0      rstudioapi_0.10  Matrix_1.2-17    tools_3.6.1
[13] glue_1.3.1       purrr_0.3.2      igraph_1.2.4.1   compiler_3.6.1
[17] pkgconfig_2.0.2  tidyselect_0.2.5 tibble_2.1.3
```
