+++ R/efs.R
@@ -1,30 +1,35 @@
-msg <- function(k, complete, val, stop_crit) {
-  cat(paste(" Edges:", k, "of", complete, "-", stop_crit, "=", round(val, 6L)),"\n")
-}
-
 #' Efficient forward selection in decomposable graphical models
 #' @description Efficient forward selection in decomposable graphical models
 #' @param df data.frame
 #' @param x A efs object
 #' @param trace Logical indicating whether or not to trace the procedure
-#' @param p Penalty term in the stopping criterion  (\code{0} = AIC and \code{1} = BIC)
-#' @param thres A threshold mechanism for choosing between two different ways of calculating the entropy. Can Speed up the procedure with the "correct" value.
-#' @return A efs object
+#' @param p Penalty term in the stopping criterion (\code{0} = AIC and \code{1} = BIC)
+#' @param thres A threshold mechanism for choosing between two different ways of calculating the entropy.
+#' @return A efs object (a list) with values:
+#' \describe{
+#' \item{\code{G_adj}}{The fitted graph. Can be extracted with \code{as_adj_lst}}
+#' \item{\code{G_A}}{The edge that was just deleted and the entropy attached to it. Can be extracted using \code{as_adj_mat}}
+#' \item{\code{CG}}{The cliques of the graph.}
+#' \item{\code{CG_A}}{A neighbor matrix of the clique graph.}
+#' \item{\code{MSI}}{Minimal separator info list}
+#' \item{\code{ht}}{An updated version of the input argument \code{ht}.}
+#' }
+#' @details The algorithm relies on calculating lots of entropies (see \url{https://arxiv.org/abs/1301.2267}). If \code{thres} is equal to a number, say \code{a}, the entropy will be calculated using the \code{table} function for all situations where the number of variables is less than or equal to \code{a}. Otherwise, the entropies will be calculated by pasting rows in the subset of \code{df} over the relevant variables. For small values of \code{a} the former method is generally faster whereas for large values of \code{a} it breaks down and it is necessary to use the latter method.
 #' @examples
 #' efs(tgp_dat[, 5:8])
 #' @references \url{https://arxiv.org/abs/1301.2267}, \url{https://doi.org/10.1109/ictai.2004.100} 
-#' @seealso \code{\link{cl_tree}}, \code{\link{efs_step}}, \code{\link{adj_list.efs}}, \code{\link{adj_matrix.efs}}
+#' @seealso \code{\link{efs_init}}, \code{\link{efs_step}}, \code{\link{bws}}, \code{\link{bws_step}}, \code{\link{cl_tree}}, \code{\link{adj_list.efs}}, \code{\link{adj_matrix.efs}}
 #' @export
 efs <- function(df, x = efs_init(df), p = 0.5, trace = TRUE, thres = 5) {
   if (!("efs" %in% class(x)) ) stop("x is not a efs class")
   lv       <- sapply(df, function(x) length(unique(x)))
   n        <- ncol(df)
-  ## FIX THIS SO IT CAN HANDLE SINGLETONS!
-  if ( n < 2 ) stop("df must have at least two variables")
-  if ( p < 0 || p > 1 ) stop("p must be between 0 and 1")
   M        <- nrow(df)
   complete <- n * (n-1L) / 2L
   k        <- sum(x$G_A)/2
+  ## FIX THIS SO IT CAN HANDLE SINGLETONS!
+  if ( n < 2 ) stop("df must have at least two variables")
[…]
