### 2019-09-01 #1665 @mlindsk
@jkanche Could you try again - I forgot to "suggest" the packages "dplyr" and "pander" in the vignette.
