### 2019-09-01 #1665 @jkanche
@mlindsk let me know if i was unclear on something

- `devtools::check()` fails, add `pander` package to either dependencies or suggests in the description file & require this package on the vignette
- I'd recommend adding a class `efs` that extends list using a proper class definition. Then create instances of this class in `efs_init` and `efs_step`.
- Also wondering if you have looked into the `tbl_graph` class from the `tidygraph` package. This would simplify a lot of things code wise and compatibility with other packages.
- in the vignette, why not use all the haplotypes instead of the first 25 on the larger example ?
- To make this reproducible, how did you process the 1000 genomes data to generate the datasets in the package ? may be add this preprocessing code as a vignette ?
- I also agree with @jdeligt, I wonder if the example dataset is good to illustrate the model
- did you validate the results from your model ?
- would a more appropriate example be metagenomic datasets ? for example, If you have taxonomic counts from samples across disease and control groups, can you use this model to find the outliers ?
