### 2019-09-01 #1665 @mlindsk
> @mlindsk let me know if i was unclear on something
>
>     * `devtools::check()` fails, add `pander` package to either dependencies or suggests in the description file & require this package on the vignette

 - Perfect you didn't have `pander` installed and cached this one. Should be fixed.

>
>     * I'd recommend adding a class `efs` that extends list using a proper class definition. Then create instances of this class in `efs_init` and `efs_step`.
>

 - Let me see if I get you right here. You want me to make a superclass (basically a constructor), `efs`, from which the subclasses `efs_init` and `efs_step` should inherit from? And also, the class of  `efs` should include the `list` class? I can make this work if this is what you are referring to as "proper class definition".

>     * Also wondering if you have looked into the `tbl_graph` class from the `tidygraph` package. This would simplify a lot of things code wise and compatibility with other packages.
>

 - I was not aware of this package; it seems interesting. What `efs` does is really just outputting an adjacency list/adjacency matrix which `tidygraph` users can easily grab and explore. I

>     * in the vignette, why not use all the haplotypes instead of the first 25 on the larger example ?
>

 - The run time of the algorithm increases and it does not add more to the example.

>     * To make this reproducible, how did you process the 1000 genomes data to generate the datasets in the package ? may be add this preprocessing code as a vignette ?
>

 - Yes; if I keep data; see below.

>     * I also agree with @jdeligt, I wonder if the example dataset is good to illustrate the model
>

 - Since you both ( @jdeligt ) I might have to find another dataset and completely remove the old one. Sure it is interesting what you mention below about metagenomic datasets. But this could easily extend to a new paper using this outlier model (somewhat overkill to explain the software). I cant stress enough that the method is *really* generic, and is a general method for doing outlier detection in contingency tables. I guess, if I asked a third person, his or her opinion would be to include a dataset from his or her respective field of research. So maybe I should really just simulate data to be neutral. But that is really not interesting. On the other hand, I could also find a handful of real world examples to cover a diverse range of applications. I am not sure what the best approach is. Don´t take this the wrong way; your comments on this are truly welcome. I'm just not quite sure of to proceed in order to make it general as explained.

>     * did you validate the results from your model ?
>

 - Validate in what sense ?

>     * would a more appropriate example be metagenomic datasets ? for example, If you have taxonomic counts from samples across disease and control groups, can you use this model to find the outliers ?

 - Yes, i believe that the model would be ideal for such a use-case since what you refer to would classify as contingency data.
