### 2019-09-02 be52593
print msg fix
