+++ R/bws_utils.R
@@ -16,77 +16,72 @@ remove_edge_mat <- function(A, e) {
 
 #' Construct a bws object
 #' @description A constructor for bws class
+#' @param df data.frame 
 #' @param adj An adjacency list
 #' @param ht An environment (hashtable) containing precomputed entropies
 #' @return A bws object. See \code{bws} for more details about the returning object.
 #' @examples
-#' d <- tgp_dat[, 5:8]
-#' bws_class(d, make_complete_graph(colnames(d)))
+#' d   <- tgp_dat[, 5:8]
+#' cg <- make_complete_graph(colnames(d))
+#' bi <- bws_init(d, cg)
+#' bws_step(d, bi)
 #' @seealso \code{\link{bws_step}}, \code{\link{bws}}, \code{\link{make_complete_graph}}
 #' @export
-bws_class <- function(df, adj, ht = new.env(hash = TRUE)) {
+bws_init <- function(df, adj, ht = new.env(hash = TRUE)) {
   stopifnot(is.environment(ht))
   stopifnot( is.list(adj) || names(adj) != length(adj))
   structure(list(
     G_adj = adj,
     G_A   = as_adj_mat(adj),
     e     = NULL,            # Best edge to delete
-    S     = character(0),    # Minimal separator for vertices in e
-    C     = list(),          # Clique list,
+    C     = rip(adj)$C,      # Clique list,
     lv    = vapply(df, function(x) length(unique(x)), 1L),
     ht    = ht),
     class = c("bws", "list")
   )
 }
 
-## Make a bws constructor and call it in bws step at the end
-## init <- function() UseMethod("init")
-## bws_class.init <- function() ...  - call this in bws
 
 #' Stepwise backward selection in decomposable graphical models
 #' @description Stepwise backward selection in decomposable graphical models
 #' @param df data.frame
-#' @param x A bws object. Can be created explicitly with \code{bws_class}
+#' @param x A bws object. Can be created explicitly with \code{bws_init}
 #' @param p Penalty term in the stopping criterion  (\code{0} = AIC and \code{1} = BIC)
 #' @param thres A threshold mechanism for choosing between two different ways of calculating the entropy. Can Speed up the procedure with the "correct" value.
 #' @return A bws object. See \code{bws} for details about the returning object.
 #' @details See \code{\link{efs}} for details about \code{thres}.
 #' @examples
-#' d <- tgp_dat[, 5:8]
-#' bws_step(d, bws_init(make_complete_graph(colnames(d))))
+#' d  <- tgp_dat[, 5:8]
+#' cg <- make_complete_graph(colnames(d))
+#' bi <- bws_init(d, cg)
+#' bws_step(d, bi)
 #' @seealso \code{\link{bws}}, \code{\link{make_complete_graph}}, \code{\link{efs}}, \code{\link{efs_step}}, \code{\link{cl_tree}}
 #' @export
 bws_step <- function(df, x, p = 0.5, thres = 5) {
   # x : bws class
-  adj     <- x$G_adj
-  ht      <- x$ht
-  cliques <- rip(adj)$C # Make a rip function that does not test for decomposability
-  A       <- x$G_A
-  nodes   <- names(adj)
+  nodes   <- names(x$G_adj)
   M       <- nrow(df)
   e_min   <- Inf
-  e_del   <- vector("character", 2L)
-  S_e     <- character(0)
-  for( i in 1:ncol(A) ) {
+  for( i in 1:ncol(x$G_A) ) {
     for( j in 1:i ) {
-      if( A[i, j] == 1L) {
[…]
