+++ man/bws_init.Rd
@@ -1,12 +1,14 @@
 % Generated by roxygen2: do not edit by hand
 % Please edit documentation in R/bws_utils.R
-\name{bws_class}
-\alias{bws_class}
+\name{bws_init}
+\alias{bws_init}
 \title{Construct a bws object}
 \usage{
-bws_class(df, adj, ht = new.env(hash = TRUE))
+bws_init(df, adj, ht = new.env(hash = TRUE))
 }
 \arguments{
+\item{df}{data.frame}
+
 \item{adj}{An adjacency list}
 
 \item{ht}{An environment (hashtable) containing precomputed entropies}
@@ -18,8 +20,10 @@ A bws object. See \code{bws} for more details about the returning object.
 A constructor for bws class
 }
 \examples{
-d <- tgp_dat[, 5:8]
-bws_class(d, make_complete_graph(colnames(d)))
+d   <- tgp_dat[, 5:8]
+cg <- make_complete_graph(colnames(d))
+bi <- bws_init(d, cg)
+bws_step(d, bi)
 }
 \seealso{
 \code{\link{bws_step}}, \code{\link{bws}}, \code{\link{make_complete_graph}}
