### 2019-09-03 #1665 @mlindsk
@csoneson, @jdeligt, @jkanche

Hi all,

Based on the comments I have now decided to make some major changes to the API. I have also implemented backward selection now, and I will try to make a much more clearer API using a common class for these methods.

According to the example data. Do you all agree that it should be wiped, and therefore cannot be modified to fully justify the outlier test? If not, how would you like it to be ideally presented then? If yes, besides @jkanche s suggestion about metagenomic data, do you have any suggestions?

@csoneson Is there a time limit I should be aware of. I'll of course strive to do things as quick as possible.
