### 2019-09-03 #1665 @mlindsk
> > @csoneson Is there a time limit I should be aware of. I'll of course strive to do things as quick as possible.
>
> No, not really. It would be great to have an estimate of the time frame, and/or if you could update us in the process.

Ok. I'll inform you concurrently.
