+++ R/efs_utils.R
@@ -26,6 +26,7 @@ is_Ca_and_Cb  <- function(m, x, y) { # m: msi object
   })
 }
 
+
 ## -----------------------------------------------------------------------------
 ##                             INITIALIZATION
 ## -----------------------------------------------------------------------------
@@ -58,6 +59,7 @@ efs_init <- function(df) { ## Should be a character matrix in the future
     x  <- pairs[[p]]
     edge_x <- sort_(x)
     ht[[edge_x]] <<- entropy(df[x])
+    ## Add the penalty here
     dst_x  <- ht[[x[1]]] + ht[[x[2]]] - ht[[edge_x]]
     if( dst_x >= max_dst ) {
       max_dst   <<- dst_x
@@ -75,109 +77,9 @@ efs_init <- function(df) { ## Should be a character matrix in the future
     MSI      = msi,
     ht       = ht
   )
-  class(out) <- c("efs")
+  class(out) <- c("efs", class(out))
   return(out)
 }
-
-## -----------------------------------------------------------------------------
-##                         STOPPING CRITERIAS
-## -----------------------------------------------------------------------------
-
-## mdl1 <- function(adj, lv, df, d, thres)  {
-##   # adj: Adjacency list
-##   # lv: not used here. Included for compatability with efs_mdl
-##   RIP      <- rip(adj)
-##   cliqs    <- RIP$C
-##   seps     <- RIP$S
-##   Nobs     <- nrow(df)
-##   Nvars    <- ncol(df)
-##   logNvars <- log(Nvars)
-##   DL_graph <- sum(sapply(cliqs, function(z) logNvars + length(z) * logNvars )) 
-##   DL_prob  <- d * sum(sapply(seq_along(cliqs), function(i) {
-##     if( i == 1L ) return( length(cliqs[[i]]) - 1 )
-##     Ci <- cliqs[[i]]
-##     Si <- seps[[i]]
-##     Ci_Si <- setdiff(Ci, Si)
-##     length(Si) * (length(Ci_Si) - 1)
-##   }))
-##   HM_C <- sum(sapply(cliqs, function(z) {
-##     dst  <- if( length(z) <= thres ) metric("entropy") else metric("entropy2")
-##     dst(df[z])
-##   }))
-##   HM_S <- 0L
-##   if( length(seps[-1]) ) {
-##     HM_S <- sum(sapply(seps[-1], function(z) {
-##       if( !neq_empt_chr(z)) return(0L)
-##       dst  <- if( length(z) <= thres ) metric("entropy") else metric("entropy2")
-##       dst(df[z])
-##     }))    
-##   }
-##   DL_data <- Nobs * (HM_C - HM_S)
-##   return( log(DL_graph + DL_prob + DL_data) )
-## }
-
-## mdl <- function(adj, lv, df, d, thres) {
-##   # adj: Adjacency list
-##   # lv:  Vector of length ncol(df) with number of levels for each var
-##   RIP    <- rip(adj)
-##   cliqs  <- RIP$C
-##   seps   <- RIP$S
-##   Nobs   <- nrow(df)
-##   Nvars  <- ncol(df)
-##   logNvars <- log(Nvars)
-##   DL_graph <- sum(sapply(cliqs, function(z) logNvars + length(z) * logNvars )) 
-##   DL_prob <- d * sum(sapply(seq_along(cliqs), function(i) {
-##     if( i == 1L ) return( prod(lv[cliqs[[i]]]) - 1 )
-##     Ci <- cliqs[[i]]
-##     Si <- seps[[i]]
[…]
