### 2019-09-03 #1665 @jdeligt
@mlindsk I would say the example can be useful to illustrate the generality of the method. I just think the genetics example would be helped by putting it in context. I.e. why it is not a population genetics tool. Why outlier detection is relevant to forensics but not to pop gen. And how this model does account for the biology while the method compared to does not.  If this context is not provided I would not be comfortable with this section or the comparison.

If you want to further demonstrate the generality maybe include another example, for me it would be a plus but not a requirement for publication.

All the best, Joep
