### 2019-09-05 cb29adb
bwd done, tree initiated
