+++ R/bws.R
@@ -1,87 +1,87 @@
-#' Backward selection in decomposable graphical models
-#' @description Backward selection in decomposable graphical models
-#' @param df data.frame
-#' @param adj An adjacency list
-#' @param p Penalty term in the stopping criterion  (\code{0} = AIC and \code{1} = BIC)
-#' @param ht An environment (hashtable) containing precomputed entropies
-#' @param trace Logical indicating whether or not to trace the procedure
-#' @param thres A threshold mechanism for choosing between two different ways of calculating the entropy. Can Speed up the procedure with the "correct" value.
-#' @return A bws object (a list) with values:
-#' \describe{
-#' \item{\code{G_adj}}{The fitted graph as an adjacency list.}
-#' \item{\code{G_A}}{The fitted graph as an adjacency matrix.}
-#' \item{\code{e}}{The edge that was just deleted and the stopping criteria attached to it.}
-#' \item{\code{C}}{A list with the (maximal) cliques of the graph.}
-#' \item{\code{ht}}{An updated version of the input argument \code{ht}.}
-#' }
-#' @details See \code{\link{efs}} for details about \code{thres}.
-#' @examples
-#' d <- tgp_dat[, 5:8]
-#' G <- make_complete_graph(colnames(d))
-#' bws(d, G)
-#' @seealso \code{\link{bws_step}}, \code{\link{make_complete_graph}}, \code{\link{efs}}, \code{\link{efs_step}}, \code{\link{cl_tree}} 
-#' @export
-bws <- function(df, adj, p = 0.5, ht = new.env(hash = TRUE), trace = TRUE, thres = 5) {
-  x  <- bws_init(df, adj, ht)
-  n  <- ncol(df)
+## #' Backward selection in decomposable graphical models
+## #' @description Backward selection in decomposable graphical models
+## #' @param df data.frame
+## #' @param adj An adjacency list
+## #' @param p Penalty term in the stopping criterion  (\code{0} = AIC and \code{1} = BIC)
+## #' @param ht An environment (hashtable) containing precomputed entropies
+## #' @param trace Logical indicating whether or not to trace the procedure
+## #' @param thres A threshold mechanism for choosing between two different ways of calculating the entropy. Can Speed up the procedure with the "correct" value.
+## #' @return A bws object (a list) with values:
+## #' \describe{
+## #' \item{\code{G_adj}}{The fitted graph as an adjacency list.}
+## #' \item{\code{G_A}}{The fitted graph as an adjacency matrix.}
+## #' \item{\code{e}}{The edge that was just deleted and the stopping criteria attached to it.}
+## #' \item{\code{C}}{A list with the (maximal) cliques of the graph.}
+## #' \item{\code{ht}}{An updated version of the input argument \code{ht}.}
+## #' }
+## #' @details See \code{\link{efs}} for details about \code{thres}.
+## #' @examples
+## #' d <- tgp_dat[, 5:8]
+## #' G <- make_complete_graph(colnames(d))
+## #' bws(d, G)
+## #' @seealso \code{\link{bws_step}}, \code{\link{make_complete_graph}}, \code{\link{efs}}, \code{\link{efs_step}}, \code{\link{cl_tree}} 
+## #' @export
+## bws <- function(df, adj, p = 0.5, ht = new.env(hash = TRUE), trace = TRUE, thres = 5) {
[…]
