+++ R/entropy.R
@@ -1,87 +1,92 @@
-na_tab <- function(df, a) {
-  ct <- table(df[, a])
-  names(dimnames(ct)) <- a # Needed for the onedimensional separators
-  ct
-}
+## na_tab <- function(df, a) {
+##   ct <- table(df[, a])
+##   names(dimnames(ct)) <- a # Needed for the onedimensional separators
+##   ct
+## }
 
-joint_entropy <- function(df) {
-  ## if( class(df) == "character" ) stop( "From entropy function: df is not a data.frame!" )
-  x  <- na_tab(df, colnames(df))
-  Nx <- sum(x)
-  entropy_table <- apply(x, seq_along(dim(x)), function(y) {
-    ifelse(y == 0 , 0, y/Nx * log(y/Nx) )
-  })
-  -sum(entropy_table)
-}
+## joint_entropy <- function(df) {
+##   ## if( class(df) == "character" ) stop( "From entropy function: df is not a data.frame!" )
+##   x  <- na_tab(df, colnames(df))
+##   Nx <- sum(x)
+##   entropy_table <- apply(x, seq_along(dim(x)), function(y) {
+##     ifelse(y == 0 , 0, y/Nx * log(y/Nx) )
+##   })
+##   -sum(entropy_table)
+## }
 
-joint_entropy2 <- function(df) {
-  A  <- apply(df, 1, paste0, collapse = "")
-  x  <- table(A)
-  Nx <- sum(x)
-  -sum(x/Nx * log(x/Nx))
-}
+## joint_entropy2 <- function(df) {
+##   A  <- apply(df, 1, paste0, collapse = "")
+##   x  <- table(A)
+##   Nx <- sum(x)
+##   -sum(x/Nx * log(x/Nx))
+## }
 
-#' Joint Entropy
-#' 
-#' @description Calculates the joint entropy over discrete variables in \code{df}
-#' 
-#' @param df data.frame
-#' @param thres A threshold mechanism for choosing between two different ways of calculating the entropy. Can Speed up the procedure with the "correct" value.
-#' @return A numeric
-#' @examples
-#' # Joint entropy over five variables in tgp_dat
-#' entropy(tgp_dat[, 5:9])
-#' 
-#' @export
-entropy <- function(df, thres = 5) {
-  if( ncol(df) <= thres ) return(joint_entropy(df))
-  else return(joint_entropy2(df))
-}
+## #' Joint Entropy
+## #' 
+## #' @description Calculates the joint entropy over discrete variables in \code{df}
+## #' 
+## #' @param df data.frame
+## #' @param thres A threshold mechanism for choosing between two different ways of calculating the entropy. Can Speed up the procedure with the "correct" value.
+## #' @return A numeric
+## #' @examples
+## #' # Joint entropy over five variables in tgp_dat
+## #' entropy(tgp_dat[, 5:9])
+## #' 
+## #' @export
+## entropy <- function(df, thres = 5) {
+##   if( ncol(df) <= thres ) return(joint_entropy(df))
+##   else return(joint_entropy2(df))
+## }
 
-entropy_difference <- function(e, S, df, ht, thres = 5) {
-  v <- unlist(es_to_vs(e))
+## entropy_difference <- function(e, S, df, ht, thres = 5) {
[…]

[…]
