### 2019-09-09 #1665 @mlindsk
@csoneson , @jdeligt , @jkanche

Update: Work is very much in progress. The new api design is being implemented. Hopefully, if everything goes as planned, I will have something ready within the next two weeks. Hopefully also sooner.
