+++ R/efs_utils.R
@@ -81,25 +81,6 @@ efs_init <- function(df) { ## Should be a character matrix in the future
   return(out)
 }
 
-delta_xic <- function(x, lv, M, p = 0.5) {
-  UseMethod("delta_xic")
-}
-
-delta_xic.efs <- function(x, lv, M, p = 0.5) {
-  # x : efs object
-  penalty     <- log(M)*p + (1 - p)*2
-  # n           <- length(lv) # ncol(df)
-  # complete    <- n * (n-1L) / 2L
-  local_info  <- x$MSI$S[[x$MSI$max$idx]]
-  e           <- local_info$e[x$MSI$max$e]
-  S           <- local_info$S
-  vs          <- es_to_vs(names(e))[[1]]
-  HM_HM_prime <- unname(e)
-  dev         <- -2*M*HM_HM_prime
-  d_parms     <- prod(lv[vs] - 1) * prod(lv[S])
-  d_aic       <- dev + penalty * d_parms
-  return(d_aic)
-}
 
 ## -----------------------------------------------------------------------------
 ##                           SUB-ROUTINES FOR efs_step
@@ -163,8 +144,36 @@ which_Cp_from_Cx_to_Cab <- function(CG_prime, C_prime_Cx, Cx, vx, Cab, Sab,  cty
   list(add = unique(add), add_tvl = unique(add_tvl)) 
 }
 
+## delta_xic <- function(x, lv, M, p = 0.5) {
+##   UseMethod("delta_xic")
+## }
+
+## delta_xic.efs <- function(x, lv, M, p = 0.5) {
+##   # x : efs object
+##   penalty     <- log(M)*p + (1 - p)*2
+##   local_info  <- x$MSI$S[[x$MSI$max$idx]]
+##   e           <- local_info$e[x$MSI$max$e]
+##   S           <- local_info$S
+##   vs          <- es_to_vs(names(e))[[1]]
+##   HM_HM_prime <- unname(e)
+##   dev         <- -2*M*HM_HM_prime               # Fix: remove "-"
+##   d_parms     <- prod(lv[vs] - 1) * prod(lv[S]) # Fix: add    "-"
+##   d_aic       <- dev + penalty * d_parms
+##   return(d_aic)
+## }
+
+stop_v <- function(x) {
+  local_info  <- x$MSI$S[[x$MSI$max$idx]]
+  local_info$e[x$MSI$max$e]  
+}
+
+## get_e.fwd <- function(x, ...)
+
 update_edges_from_C_primes_to_Cab <- function(df, Cps, Cab, va, vb, ht, thres = 5) {
   # Cps : C_primes
+  M <- nrow(df)
+  q <- 0
+  LV  <- sapply(df, function(s) length(unique(s)))
   sep <- lapply(Cps, function(Cp) {
     Sp          <- intersect(Cp, Cab)
     eligs_Cab   <- setdiff(Cab, Cp)
@@ -174,45 +183,22 @@ update_edges_from_C_primes_to_Cab <- function(df, Cps, Cab, va, vb, ht, thres = 
     eligs_names <- eligs
     H_Sp        <- 0L
     if( neq_empt_chr(Sp) ) H_Sp <- ht[[sort_(Sp)]]
-    
-    eligs  <- sapply(eligs, function(e) {
-      ## See the proof of Theorem 4.3 in Jordan to optimize! (Dont need to use exists for all cases)
-      v <- unlist(es_to_vs(e))
-      H_Sp_x <- 0L        
-      Spx <- sort_(c(Sp, v[1]))
-      if( exists(Spx, envir = ht, inherits = FALSE) ) {
-        H_Sp_x <- ht[[Spx]]
[…]
