+++ R/helper_step_fwd.R
@@ -1,0 +1,144 @@
+fwd_init <- function(x, df, q = 0.5) {
+  # x : fwd object
+  M     <- nrow(df)
+  nodes <- colnames(df)
+  n     <- length(nodes)
+  pairs <- utils::combn(nodes, 2,  simplify = FALSE)
+  for( j in 1:n ) x$MEM[[nodes[j]]] <- entropy(df[nodes[j]])
+  x$MSI <- lapply(seq_along(pairs), function(p) {
+    v      <- pairs[[p]]
+    edge_v <- sort_(v)
+    x$MEM[[edge_v]] <<- entropy(df[v])
+    ed       <- x$MEM[[v[1]]] + x$MEM[[v[2]]] - x$MEM[[edge_v]]
+    penalty  <- log(M)*q + (1 - q)*2
+    dev      <- 2 * M * ed
+    d_parms  <- -prod(x$LV[v] - 1)
+    d_qic    <- dev + penalty * d_parms
+    if( d_qic >= attr(x$e, "d_qic") ) {
+      x$e <<- new_edge(edge_v, d_qic, p)
+    }
+    list(C1 = v[1], C2 = v[2], S = character(0L), e = structure(ed, names = edge_v))
+  })
+  attr(x$e, "ins") <- match(es_to_vs(x$e)[[1]], x$CG)
+  return(x)
+}
+
+find_new_edge <- function(msi_prime, CG_prime) {
+  max_es   <- sapply(msi_prime, function(z) z$e[which.max(z$e)])
+  max_idx  <- which.max(max_es)
+  max_e    <- names(max_idx)
+  max_atom <- msi_prime[[max_idx]]
+  max_ins  <- unlist(sapply(seq_along(CG_prime), function(x) {
+    cond <- setequal(CG_prime[[x]], max_atom$C1) || setequal(CG_prime[[x]], max_atom$C2)
+    if( cond ) return(x)
+  }))
+  new_edge(max_e, unname(max_es[max_idx]), max_idx, max_ins)
+}
+
+# m : msi object
+# Cx: clique (vector of characters) 
+is_Cx <- function(m, Cx) sapply(m, function(x) setequal(x$C1, Cx) || setequal(x$C2, Cx))
+
+is_Ca_or_Cb <- function(m, x, y) {
+  sapply(m, function(z) {
+    is_CaCb <- setequal(z$C1, x) || setequal(z$C2, y)
+    is_CbCa <- setequal(z$C1, y) || setequal(z$C2, x)
+    is_CaCb || is_CbCa
+  })
+}
+
+is_Ca_and_Cb <- function(m, x, y) {
+  sapply(m, function(z) {
+    is_CaCb <- setequal(z$C1, x) && setequal(z$C2, y)
+    is_CbCa <- setequal(z$C1, y) && setequal(z$C2, x)
+    is_CaCb || is_CbCa
+  })
+}
+
+make_G_dbl_prime <- function(Sab, G_A) {
+  keepers <- setdiff(dimnames(G_A)[[1]], Sab)
+  G_A[keepers, keepers]
+}
+
+edges_to_delete <- function(prone_to_deletion, TVL, MSab, Cab, Sab, cta, ctb) {
+  ## TVL  : Temporary vertex list
+  ## MSab : logical indicating which prone_to_deletion (those with C1 \cap C2 = Sab) that has eab = (va, vb)
+  etd <- sapply(seq_along(MSab), function(k) {
+    delete <- FALSE
+    x <- prone_to_deletion[[k]]
+    if( MSab[k] ) {
+      if( !all(x$C1 %in% Cab) && !all(x$C2 %in% Cab) ) {
[…]
