+++ R/helper_outlier.R
@@ -1,0 +1,240 @@
+## ---------------------------------------------------------
+##                NON-EXPORTED HELPERS
+## ---------------------------------------------------------
+new_outlier_model <- function(A, sims, mu, sigma, cdf, Cms, Sms) {
+  structure(
+    list(
+      A     = A,
+      sims  = sims,
+      mu    = mu,
+      sigma = sigma,
+      cdf   = cdf,
+      Cms   = Cms,
+      Sms   = Sms
+    ),
+    class = c("outlier_model", "list")
+  )
+}
+
+new_outlier <- function(m, dev, pv, g) {
+  # m : outlier_model object
+  m$dev  <- dev
+  m$pval <- pv
+  m$g    <- g
+  class(m) <- c("outlier", class(m))
+  return(m)
+}
+
+## Shut up CRAN check - foreach and z are good friends!
+utils::globalVariables('z') 
+
+.sim_internal <- function(A,
+                  C_marginals,
+                  S_marginals,
+                  nsim         = 1000,
+                  type         = "lr",
+                  ncores       = 1) {
+  y <- replicate(nsim, vector("character", ncol(A)), simplify = FALSE)
+  M <- nrow(A)
+  Delta   <- colnames(A)
+  C1_vars <- attr(C_marginals[[1]], "vars")
+  C1_idx  <- match(C1_vars, Delta)
+  p_nC1   <- C_marginals[[1]] / M
+  yC1_sim <- sample(names(p_nC1), nsim, replace = TRUE, prob = p_nC1)
+  if(!( length(C_marginals) - 1L)) {
+    # The complete graph
+    yC1_sim <- lapply(strsplit(yC1_sim, ""), function(z) {names(z) = C1_vars; z})
+    return( sapply(yC1_sim, TY, C_marginals, S_marginals) )
+  }
+  doParallel::registerDoParallel(ncores)
+  combine_ <- switch(type, "deviance"  = 'c', "raw" = "rbind")
+  y <- foreach::`%dopar%`(foreach::foreach(z = 1:nsim, .combine = combine_, .inorder = FALSE), {
+    y_sim_z <- y[[z]]
+    y_sim_z[C1_idx] <- .split_chars(yC1_sim[1])
+    for( k in 2:length(C_marginals) ) {
+      nCk     <- C_marginals[[k]]
+      Ck_vars <- attr(nCk, "vars")     # Clique names
+      Ck_idx  <- match(Ck_vars, Delta) # Where is Ck in Delta
+      nSk     <- S_marginals[[k]]      # For Sk = Ø we have that nSk = M
+      Sk_vars <- attr(nSk, "vars")     # Separator names
+      if( is.null(Sk_vars) ) {         # For empty separators
+        p_nCk_minus_nSk <- nCk / nSk   # nSk = M !
+        y_sim_z[Ck_idx] <- .split_chars(sample(names(p_nCk_minus_nSk), 1L, prob = p_nCk_minus_nSk))
+      } else {
+        Sk_idx              <- match(Sk_vars, Delta)
+        Sk_idx_in_Ck        <- match(Sk_vars, Ck_vars)
+        Ck_idx_minus_Sk_idx <- Ck_idx[-Sk_idx_in_Ck]
+        ySk                 <- y_sim_z[Sk_idx]
+        nSk_ySk             <- na_ya(nSk, paste0(ySk, collapse = ""))
+        nCk_given_Sk        <- n_b(nCk, structure(Sk_idx_in_Ck, names = ySk) )
[…]
