+++ R/api_fit_graph.R
@@ -9,13 +9,13 @@
 #' @param wrap logical specifying if the result of a run with type = "tree" should be converted to a "fwd" object
 #' @return A \code{gengraph} object with values:
 #' @examples
-#' d <- tgp_dat[, 5:8]
+#' d <- tgp_dat[1:100, 5:8]
 #' g <- fit_graph(d)
 #' # plot(g)
 #' # adj_mat(g)
 #' # adj_lst(g)
 #' @references \url{https://arxiv.org/abs/1301.2267}, \url{https://doi.org/10.1109/ictai.2004.100} 
-#' @seealso \code{\link{adj_lst.efs}}, \code{\link{adj_mat.efs}}, \code{\link{step.fwd}}, \code{\link{step.bwd}}, \code{\link{gengraph}}
+#' @seealso \code{\link{adj_lst.gengraph}}, \code{\link{adj_mat.gengraph}}, \code{\link{step.fwd}}, \code{\link{step.bwd}}, \code{\link{gengraph}}
 #' @export
 fit_graph <- function(df,
                       type  = "fwd", # (fwd, bwd, tree)
