### 2019-09-23 #1665 @csoneson
Hi @mlindsk - just checking in here. Could you give us an update on the progress of the revision?
