### 2019-09-23 #1665 @mlindsk
Hi @csoneson - The new API is now finished. I need to make a new vignette and some changes to the paper. I'm a little swamped regarding teaching, but hopefully it wont be long until I'm done now. I have merged the new code to the master branch.
