### 2019-09-26 d1672f5
api makeover and new examples
