+++ paper/paper.bib
@@ -1,9 +1,10 @@
-@article{outlier_detection_in_contingency_tables,
-  title={Outlier Detection in Contingency Tables using Decomposable Graphical Models},
-  author={Lindskou, Mads and Eriksen, Poul Svante and Tvedebrink, Torben},
+@article{lindskououtlier,
+  title={Outlier Detection in Contingency Tables Using Decomposable Graphical Models},
+  author={Lindskou, Mads and Svante Eriksen, Poul and Tvedebrink, Torben},
   journal={Scandinavian Journal of Statistics},
+  publisher={Wiley Online Library},
   doi={10.1111/sjos.12407},
-  year=2019
+  year={2019}
 }
 
 @inproceedings{deshpande01_efficient,
