+++ paper/paper.md
@@ -26,7 +26,7 @@ output: pdf_document
 
 # Summary
 
-Outlier detection is an important task in statistical analyses. An outlier is a case-specific unit since it may be interpreted as natural extreme noise in some applications, whereas in other applications it may be the most interesting observation. The **molic** package have been written to facilitate the novel outlier detection method in high-dimensional contingency tables [@outlier_detection_in_contingency_tables]. In other words, the method works for data sets in which all variables are _categorical_, implying that they can only take on a finite set of values (also called _levels_).
+Outlier detection is an important task in statistical analyses. An outlier is a case-specific unit since it may be interpreted as natural extreme noise in some applications, whereas in other applications it may be the most interesting observation. The **molic** package have been written to facilitate the novel outlier detection method in high-dimensional contingency tables [@lindskououtlier]. In other words, the method works for data sets in which all variables are _categorical_, implying that they can only take on a finite set of values (also called _levels_).
 
 The software uses decomposable graphical models (DGMs), where the probability mass function can be associated with an interaction graph, from which conditional independences among the variables can be inferred. This gives a way to investigate the underlying nature of outliers. This is also called _understandability_ in the literature. Outlier detection has many applications including areas such as
 
@@ -34,6 +34,7 @@ The software uses decomposable graphical models (DGMs), where the probability ma
  - Medical and public health
  - Anomaly detection in text data
  - Fault detection (on critical systems) 
+ - Forensic Science
 
 # The Method
 
@@ -43,7 +44,9 @@ The method can be described by the **outlier test** procedure below. Assume we a
 
 ![](outlier_test_alg.pdf)\
 
-The `efs` algorithm is an implementation of the efficient step-wise selection procedure [@deshpande01_efficient] used for model selection in decomposable graphs. This function can be used to explore dependencies between any kind of discrete variables and make statements about conditional dependencies and independencies. A thorough description of the outlier detection method and how to use the software can be found at 
+The `fit_graph` algorithm has three ways of fitting a graph. The `fwd` type (which is default) is an implementation of the efficient step-wise selection procedure [@deshpande01_efficient] used for model selection in decomposable graphs. There is also a backward, `bwd`, type and finally it is also possible to fit a tree interaction graph, i.e. only first order associations.
+
+The `fit_graph` function can be used to explore dependencies between any kind of discrete variables and make statements about conditional dependencies and independencies. A thorough description of the outlier detection method and how to use the software can be found at 
 
 \begin{center}
 https://mlindsk.github.io/molic/
@@ -52,17 +55,17 @@ https://mlindsk.github.io/molic/
 # Expert Knowledge
 If one has prior knowledge of the underlying nature of the association between variables, this can easily be exploited. One can choose to model only the relationship between variables which have no other associations to any of the remaining variables. This will result in a number of interaction graphs which can then be unified as the union of these graphs. This approach was taken in the example below.
 
-# A Use Case in Genetics
+# A Use Case in Forensic Science
 
-Recently, advances in DNA sequencing has made it possible to sequence short segments of DNA ($< 200$ basepairs) including two or more SNPs. These are called \textsl{microhaplotypes} (or microhaps for short) [@kidd2014microhaplo]. They have been demonstrated to be well suited for ancestry assessment. The short distance between SNPs within a microhap implies that recombination among them rarely occurs. Hence, the methodology of @tvedebrink2018aims can not be used as this assumes mutual independence of the SNPs within a population.
[…]
