### 2019-09-26 #1665 @mlindsk
@jkanche I will provide the preprocessing code of 1000GP soon
@jdeligt I hope I have accomodated your suggestion about generality/forensics/pop gen. According to DOIs; I don't see any DOI for "Efficient stepwise selection in decomposable models" which is why I still haven't included it (did I miss something?).
