### 2019-09-26 #1665 @jdeligt
@mlindsk thanks for the update and considering our suggestions. I'm happy with how you've now framed the example and the statement of need. DOI's are indeed in order, sorry for confusion. I'm happy to sign off on it.
