### 2019-09-27 #1665 @mlindsk
@jkanche Preprocessing script is now included along with the original 1000GP data
