### 2019-10-01 #1665 @mlindsk
Good morning all @jdeligt @jkanche @csoneson

It seems that we are in a final stage. @jkanche could you check for installation? I believe this is the only checkbox left. @csoneson would you consider changing the version number to v.1.0.0?
