### 2019-10-01 2f453cc
Paper tweaks
