### 2019-10-01 #1665 @csoneson
@jdeligt, @jkanche - thanks for your reviews! And thanks @jdeligt for letting us know that you are happy with the current state.

@mlindsk - I have sent a PR with some fixes to the references in the paper, and a couple of additional comments.
The version number will be set once the submission is accepted and archived in Zenodo (it will not change in this issue).
