### 2019-10-01 #1665 @mlindsk
> @mlindsk - I have sent a PR with some fixes to the references in the paper, and a couple of additional comments.

The saturated model is now motivated - it is a natural "go to model", but I argue that one shouldn't use it.
