### 2019-10-02 f8ae47c
data.frames of single columns can be used in fit_graph now
