+++ README.Rmd
@@ -53,6 +53,39 @@ devtools::install_github("mlindsk/molic", build_vignettes = FALSE)
 TBA
 ```
 
+## Main Functions
+The main functions in **molic** are
+
+ - `fit_graph` which fits a decomposable graph. It has three types; forward selection (`fwd`), backward selection (`bwd`) and tree (`tree`). Using `adj_lst` on an object returned by `fit_graph` gives the **adjacency list** corresponding to the graph. Similarly one can use `adj_mat` to obtain an adjacency matrix.
+ - `fit_outlier` which can be used to test if an observation is an outlier in some categorical data. It needs an adjacency list as input which can be obtained from an object returned by `fit_graph`.
+ 
+Adjacency lists are important in **molic**. They are named `list` objects of the form
+
+```{r}
+adj <- list(a = "b", b = c("a", "c"), c = "b", d = character(0))
+```
+
+We can plot the corresponding graph by creating a `gengraph` object as
+
+
+```{r echo = FALSE, message = FALSE, warning = FALSE}
+library(molic)
+```
+
+```{r  gengraph, fig.align = "center"}
+d <- data.frame(a = "", b = "", c = "", d = "") # A dummy dataframe
+g <- gengraph(d, type = "gen", adj)
+plot(g)
+```
+The dummy `data.frame` `d` is needed as an argument. This is because, in almost all cases a `gengraph` object is obtained from `fit_graph` which uses a `data.frame` to fit the graph. Notice, that **isolated** nodes (here `d`) is formed using the empty character `character(0)`.
+
+Finally, since the `fit_outlier` function assumes that `adj` is **decomposable** (the graph cannot have cycles of length greater than $4$ without a chord) we can test for this explicitly
+
+```{r}
+is_decomposable(adj)
+```
+This can be useful, if the user has obtained an adjacency list using some other software than `fit_graph`. The `fit_outlier` will also raise a warning if the graph (`adj`) is not decomposable.
+
 ## Example - Outlier Detection
 
 To demonstrate the outlier method we use the `car` data set from the [UCI Machine Learning Repository](https://archive.ics.uci.edu/ml/index.php). The data have $4$ classes that labels the evaluation of a car; `unacceptable, acc, vgood` and `good`. These classes are determined by the other variables in the data - and theses are _not_ necessarily independent of each other and we must therefore "fit their association".
@@ -81,10 +114,6 @@ unacc_cars <- car %>%
 ```
 
 ### Fitting an Interaction Graph
-
-```{r echo = FALSE, message = FALSE, warning = FALSE}
-library(molic)
-```
 
 We fit the interaction graph for the `vgood` cars and plot the result.
 
