+++ Makefile
@@ -25,6 +25,8 @@ test:
 
 readme:
 	Rscript -e "rmarkdown::render('README.Rmd')"; \
+	convert man/figures/README-gengraph-1.png -gravity south -chop 0x100 man/figures/README-gengraph-1.png; \
+	convert man/figures/README-gengraph-1.png -gravity north -chop 0x80 man/figures/README-gengraph-1.png; \
 	convert man/figures/README-acc-1.png -gravity south -chop 0x100 man/figures/README-acc-1.png; \
 	convert man/figures/README-acc-1.png -gravity north -chop 0x80 man/figures/README-acc-1.png; \
 	convert man/figures/README-unacc-1.png -gravity south -chop 0x100 man/figures/README-unacc-1.png; \
