### 2019-10-07 #1665 @mlindsk
@jkanche Is it possible for you to make the installation check? I believe this is one of last things in the process now.
