### 2019-10-07 #1665 @csoneson
@mlindsk - I noticed one more typo in the paper: first sentence of the use case should be "advances [...] _have_", not "advances [...] has". Could you fix this and generate a new pdf proof (`@whedon generate pdf`)?

The next steps are then as follows:
- [ ] Check that the software version is consistently reported
- [ ] Create a tagged release of the current version of the software and report the version tag here
- [ ] Archive the software in Zenodo. Make sure that the title and authors of the archive are the same as those of the paper (you can also add ORCID)
- [ ] Report back the doi of the Zenodo archive here
