### 2019-10-08 #1665 @mlindsk
@csoneson

 - Tag: v1.0.0
 - DOI: 10.5281/zenodo.3475854
