### 2019-10-08 #1665 @mlindsk
> @mlindsk Please fix the metadata of the Zenodo archive - the title and authors should be the same as the paper.

@csoneson I was a little too fast I see; should be fixed now.
