### 2019-10-08 #1665 @csoneson
@mlindsk - ok, looks good! I'll hand over to @openjournals/joss-eics to finalize the process
