### 2019-10-09 #1665 @labarba
The acronym SNP is used without ever spelling it out. (Of course, also DNA, but this being of popular knowledge, it can pass.)
