### 2019-10-09 #1665 @labarba
@mlindsk 👋  — I am the Associate Editor-in-Chief on rotation this week. Will you add the meaning of the acronym SNP to the paper before we publish?
