### 2019-10-10 #1665 @mlindsk
Hi @labarba
The meaning of the SNP acronym is now added.
