### 2019-10-10 #1665 @labarba
Huge thanks to our editor: @csoneson, and the reviewers: @jdeligt, @jkanche — your contribution to JOSS is greatly appreciated! 🙏
